requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
Pillow==10.0.0
numpy==1.24.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
bcrypt==4.0.1
blinker==1.6.3
greenlet==2.0.2
SQLAlchemy==2.0.23
typing_extensions==4.8.0